Added services for entities

+++ api/src/main/java/com/marcinolek/mytimesheet/controller/base/AbstractCrudController.java
@@ -1,5 +1,6 @@
 package com.marcinolek.mytimesheet.controller.base;
 
+import com.marcinolek.mytimesheet.controller.auth.AuthController;
 import com.marcinolek.mytimesheet.dto.base.AbstractDTO;
 import com.marcinolek.mytimesheet.dto.pagination.PaginationRequestDTO;
 import com.marcinolek.mytimesheet.dto.pagination.PaginationResponseDTO;
@@ -15,7 +16,7 @@ import java.io.Serializable;
 import java.util.List;
 
 public abstract class AbstractCrudController<TEntity extends AbstractEntity, TDto extends AbstractDTO, ID extends Serializable>
-        extends BaseController {
+        extends AuthController {
 
     @Autowired
     private AbstractCrudService<TEntity, TDto, ID> service;
@@ -29,31 +30,43 @@ public abstract class AbstractCrudController<TEntity extends AbstractEntity, TDt
     }
 
     @PostMapping("/query")
-    public PaginationResponseDTO<TDto> getAll(@RequestBody PaginationRequestDTO request) {
+    public PaginationResponseDTO<TDto> getAll(@RequestBody PaginationRequestDTO request) throws WebApiException {
+        this.checkReadPermission();
         Page<TDto> result = this.service.findAll(request);
         return new PaginationResponseDTO<>(result);
     }
 
     @GetMapping("/{id}")
     public TDto getById(@PathVariable ID id) throws WebApiException {
+        this.checkReadPermission();
         return this.service.findById(id);
     }
 
     @PutMapping("/{id}")
     public TDto update(@PathVariable ID id, @RequestBody TDto dto) throws WebApiException {
+        this.checkUpdatePermission();
         return this.service.update(dto, id);
     }
 
     @PostMapping("")
     public TDto create(@RequestBody TDto dto) throws WebApiException {
+        this.checkCreatePermission();
         return this.service.create(dto);
     }
 
     @DeleteMapping("/{id}")
     @ResponseBody
     public void deleteById(@PathVariable ID id) throws WebApiException {
+        this.checkDeletePermission();
         this.service.deleteById(id);
     }
 
+    protected abstract void checkCreatePermission() throws WebApiException;
+
+    protected abstract void checkReadPermission() throws WebApiException;
+
+    protected abstract void checkUpdatePermission() throws WebApiException;
+
+    protected abstract void checkDeletePermission() throws WebApiException;
 
 }
